Refactor include guards and forward declarations to fix circular dependency between Public Services and PublicServicesState Renamed file "Public Services.h" and "Public Services.cpp" to "PublicServices.h" and "PublicServices.cpp" respectively. Added default constructors to the Educator, Modern,Outdated and Standard classes, but did not implement them, Educator constructor currently only initialises its attributes to null.
