+++ src/Government/Education.h
@@ -1,12 +1,13 @@
 #ifndef EDUCATION_H
 #define EDUCATION_H
 
-#include "Public Services.h"
+#include "PublicServices.h"
 
 class Education : public Public_Services {
 	private:
 		PublicServiceState* InfraStructureCondition;
 	public:
+	Education();
     std::string getTypeOfPublicService();
     void increaseBudget();
 		void decreaseBudget();
