+++ src/Government/HealthCare.h
@@ -1,7 +1,7 @@
 #ifndef HEALTHCARE_H
 #define HEALTHCARE_H
 
-#include "Public Services.h"
+#include "PublicServices.h"
 
 class HealthCare : public Public_Services {
 	private:
