+++ src/Government/Outdated.h
@@ -4,10 +4,12 @@
 #include "PublicServiceState.h"
 #include "Standard.h"
 
+class Standard;
 class Outdated : public PublicServiceState {
 	private:
 		double OutdatedHighCap = 100;
 	public:
+		Outdated();
 		void increaseBudget(double increase);
 		void descreaseBudget(double decrease);
 		std::string getType();
