+++ src/Government/Standard.h
@@ -5,11 +5,14 @@
 #include "Modern.h"
 #include "Outdated.h"
 
+class Modern;
+class Outdated;
 class Standard : public PublicServiceState {
 	private:
 		double StandardHighCap = 1000;
 		double StandardLowerCap = 100;
 	public:
+		Standard();
 		void increaseBudget(double increase);
 		void descreaseBudget(double decrease);
 		std::string getType();
