+++ src/Makefile
@@ -53,6 +53,17 @@ SRCS = $(SRC_DIR)/main.cpp \
        $(SRC_DIR)/buildingFactories/ResidentialBuildingFactory.cpp \
        $(SRC_DIR)/buildingFactories/UtilityBuildingFactory.cpp \
        $(SRC_DIR)/buildingFactories/TransportBuildingFactory.cpp \
+       $(SRC_DIR)/Government/Standard.cpp \
+       $(SRC_DIR)/Government/Outdated.cpp \
+       $(SRC_DIR)/Government/Modern.cpp \
+       $(SRC_DIR)/Government/PublicServices.cpp \
+       $(SRC_DIR)/Government/CitySubject.cpp \
+       $(SRC_DIR)/Government/Education.cpp \
+       $(SRC_DIR)/Government/HealthCare.cpp \
+       $(SRC_DIR)/Government/Police.cpp \
+       $(SRC_DIR)/Government/BoostHealthCarePolicy.cpp \
+       $(SRC_DIR)/Government/BoostPolicePolicy.cpp \
+       $(SRC_DIR)/Government/BoostEducationPolicy.cpp \
 
 
 # Object files
